chore(list-view): naversList componentized
