+++ src/components/Navers/index.ts
@@ -1,0 +1,1 @@
+export { default as NaversListView } from './NaversListView';
