+++ src/components/index.ts
@@ -6,3 +6,4 @@ export * from './Modal';
 export * from './Input';
 export * from './Button';
 export * from './Drawer';
+export * from './Navers';
